optimize evaluation on useSplitTreatments using a memoized version of client.getTreatmentsWithConfig
