+++ src/useSplitTreatments.ts
@@ -1,5 +1,6 @@
+import React from 'react';
 import { getControlTreatmentsWithConfig, ERROR_UT_NO_USECONTEXT } from './constants';
-import { checkHooks, IClientWithContext } from './utils';
+import { checkHooks, IClientWithContext, memoizeGetTreatmentsWithConfig } from './utils';
 import { ISplitTreatmentsChildProps, IUpdateProps } from './types';
 import { INITIAL_CONTEXT } from './SplitContext';
 import { useSplitClient } from './useSplitClient';
@@ -14,8 +15,11 @@ import { useSplitClient } from './useSplitClient';
 export function useSplitTreatments(splitNames: string[], attributes?: SplitIO.Attributes, key?: SplitIO.SplitKey, options?: IUpdateProps): ISplitTreatmentsChildProps {
   const context = checkHooks(ERROR_UT_NO_USECONTEXT) ? useSplitClient(key, undefined, undefined, options) : INITIAL_CONTEXT;
   const client = context.client;
+
+  const getTreatmentsWithConfig = React.useMemo(memoizeGetTreatmentsWithConfig, []);
+
   const treatments = client && (client as IClientWithContext).__getStatus().isOperational ?
-    client.getTreatmentsWithConfig(splitNames, attributes) :
+    getTreatmentsWithConfig(client, context.lastUpdate, splitNames, attributes, { ...client.getAttributes() }) :
     getControlTreatmentsWithConfig(splitNames);
 
   return {
